Add files via upload
Automatic Fading ON and OFF Circuit using BC847, BC557, Capacitor and Resistor

 

When energized with +4V in the circuit I designed, the leds will flash at a certain speed and go out again at the same speed. Of course, this speed is quite slow. You can make beautiful lighting for your home by establishing a circuit similar to this.

 

BC847 

 

BC847 is in SMD Transistor Series and in SOT-23 sheath. BC847 is an SMD NPN type Transistor. The BC847 Transistor has a collector current of 100mA and a Collector-Emitter voltage of 45V.

 

Components I use:

-5mm Led x5

-BC847

-BC557 X2

-33k, 10k, 12k, 2.2k Resistor

-470uF, 100uF Capasitor